refactor(rich-ui): split rich_ui_controller.rb into modular directory structure (Phase 1)
Extract the 2336-line lib/clacky/rich_ui_controller.rb into focused modules:

- lib/clacky/rich_ui.rb                 # new entry point
- lib/clacky/rich_ui/rich_ui_controller.rb # thin controller
- lib/clacky/rich_ui/shell/rich_agent_shell.rb
- lib/clacky/rich_ui/components/
  - sidebar.rb, sidebar_panels.rb
  - thinking_live_view.rb, status_view.rb
  - dialogs/config_menu_dialog.rb
  - dialogs/form_dialog.rb
  - dialogs/approval_dialog.rb
- lib/clacky/rich_ui/layout_adapter.rb
- lib/clacky/rich_ui/progress_handle_adapter.rb
- lib/clacky/rich_ui/extensions/ruby_rich_patches.rb

Other changes:
- Update CLI to require_relative 'rich_ui' instead of 'rich_ui_controller'
- Keep lib/clacky/rich_ui_controller.rb as a compatibility shim
- Fix RubyRich::Theme.whale_dark -> agent_dark for ruby_rich 0.4.8
- Add docs/rich_ui_refactor_plan.md
