+++ lib/clacky/rich_ui/layout_adapter.rb
@@ -1,0 +1,14 @@
+# frozen_string_literal: true
+
+module Clacky
+    class LayoutAdapter
+      def initialize(shell)
+        @shell = shell
+      end
+
+      def clear_output
+        @shell.transcript.store.entries.clear
+        @shell.viewport.scroll_to_bottom
+      end
+    end
+end
